refactor: prepare parameter specification for additional baysian optimization config
